clean up / refactoring
