+++ .gitignore
@@ -45,3 +45,8 @@ app/services/__pycache__/league_service.cpython-312.pyc
 app/routes/__pycache__/league_routes.cpython-312.pyc
 app/routes/__pycache__/league_routes.cpython-312.pyc
 app/routes/__pycache__/league_routes.cpython-312.pyc
+app/routes/__pycache__/league_routes.cpython-310.pyc
+app/services/__pycache__/league_service.cpython-310.pyc
+business_logic/__pycache__/definitions.cpython-310.pyc
+data_access/__pycache__/schema.cpython-310.pyc
+database/generator/__pycache__/seed.cpython-310.pyc
