+++ 99_deprecated/app.py
@@ -10,7 +10,7 @@ import os
 
 #sys.path.append(os.path.dirname(os.path.dirname(__file__)))
 from data_access.read import read
-from database.definitions import Columns
+from data_access.schema import Columns
 from business_logic.statistics import calculate_score_average_player, calculate_score_min_player, calculate_score_max_player, calculate_games_count_player, calculate_score_average_league
 app = Flask(__name__, template_folder='html/templates')
