--- database/data/tmp.csv
@@ -1,25 +1,0 @@
-Season;Week;Datum;League Name;Spielort;Team Name;Name;EDV;Match Number;Team Name Opponent;Position;Pins
-24/25;5;n/a;LL1 Nord;Augsburg;Augsburg 70;Ben Moore;98407;4;Neu-Ulm 27;0;176
-24/25;5;n/a;LL1 Nord;Augsburg;Augsburg 70;Finn Friedrich;51767;4;Neu-Ulm 27;1;161
-24/25;5;n/a;LL1 Nord;Augsburg;Augsburg 70;Emil Hahn;88813;4;Neu-Ulm 27;2;215
-24/25;5;n/a;LL1 Nord;Augsburg;Augsburg 70;Robert Schäfer;78955;4;Neu-Ulm 27;3;227
-24/25;5;n/a;LL1 Nord;Augsburg;Fürth 26;Diego Becker;77954;4;Kempten 61;0;143
-24/25;5;n/a;LL1 Nord;Augsburg;Fürth 26;Patrick Weiß;51772;4;Kempten 61;1;210
-24/25;5;n/a;LL1 Nord;Augsburg;Fürth 26;Moritz Rossi;64908;4;Kempten 61;2;204
-24/25;5;n/a;LL1 Nord;Augsburg;Fürth 26;Julian Lehmann;22275;4;Kempten 61;3;133
-24/25;5;n/a;LL1 Nord;Augsburg;Kempten 61;Felix Bergmann;40627;4;Fürth 26;0;200
-24/25;5;n/a;LL1 Nord;Augsburg;Kempten 61;Michael Bauer;28765;4;Fürth 26;1;170
-24/25;5;n/a;LL1 Nord;Augsburg;Kempten 61;Mohammad Krause;55333;4;Fürth 26;2;239
-24/25;5;n/a;LL1 Nord;Augsburg;Kempten 61;Daniel Schulz;85916;4;Fürth 26;3;164
-24/25;5;n/a;LL1 Nord;Augsburg;Neu-Ulm 27;Linus Krause;13646;4;Augsburg 70;0;191
-24/25;5;n/a;LL1 Nord;Augsburg;Neu-Ulm 27;Heinz Köhler;79030;4;Augsburg 70;1;193
-24/25;5;n/a;LL1 Nord;Augsburg;Neu-Ulm 27;Vincent Zimmermann;92698;4;Augsburg 70;2;169
-24/25;5;n/a;LL1 Nord;Augsburg;Neu-Ulm 27;Florian Brown;39806;4;Augsburg 70;3;160
-24/25;5;n/a;LL1 Nord;Augsburg;Neu-Ulm 48;Ben Richter;94424;4;Würzburg 40;0;226
-24/25;5;n/a;LL1 Nord;Augsburg;Neu-Ulm 48;William Sousa;93180;4;Würzburg 40;1;204
-24/25;5;n/a;LL1 Nord;Augsburg;Neu-Ulm 48;Jackson Nguyen;62138;4;Würzburg 40;2;211
-24/25;5;n/a;LL1 Nord;Augsburg;Neu-Ulm 48;Niklas Riedel;35365;4;Würzburg 40;3;201
-24/25;5;n/a;LL1 Nord;Augsburg;Würzburg 40;Tobias Koch;69235;4;Neu-Ulm 48;0;206
-24/25;5;n/a;LL1 Nord;Augsburg;Würzburg 40;Kevin Frank;28037;4;Neu-Ulm 48;1;165
-24/25;5;n/a;LL1 Nord;Augsburg;Würzburg 40;Lars Schäfer;25895;4;Neu-Ulm 48;2;228
-24/25;5;n/a;LL1 Nord;Augsburg;Würzburg 40;Kilian Busch;19336;4;Neu-Ulm 48;3;157
