--- tmp.csv
@@ -1,25 +1,0 @@
-Season;Week;Date;League;Location;Team;Player;Player ID;Match Number;Opponent;Position;Score;Points
-24/25;5;n/a;BZL 2 Nord;Passau;Hof 79;Ben Dietrich;77062;4;Schweinfurt 23;0;144;
-24/25;5;n/a;BZL 2 Nord;Passau;Hof 79;Ben Dietrich;77062;4;Passau 51;0;155;
-24/25;5;n/a;BZL 2 Nord;Passau;Hof 79;Martin Werner;11590;4;Schweinfurt 23;1;132;
-24/25;5;n/a;BZL 2 Nord;Passau;Hof 79;Martin Werner;11590;4;Passau 51;1;127;
-24/25;5;n/a;BZL 2 Nord;Passau;Hof 79;Florian Friedrich;50625;4;Schweinfurt 23;2;126;
-24/25;5;n/a;BZL 2 Nord;Passau;Hof 79;Florian Friedrich;50625;4;Passau 51;2;138;
-24/25;5;n/a;BZL 2 Nord;Passau;Hof 79;Oliver Schmitt;47804;4;Schweinfurt 23;3;114;
-24/25;5;n/a;BZL 2 Nord;Passau;Hof 79;Oliver Schmitt;47804;4;Passau 51;3;147;
-24/25;5;n/a;BZL 2 Nord;Passau;Ingolstadt 78;Matthias Thomas;91252;4;Kempten 57;0;178;
-24/25;5;n/a;BZL 2 Nord;Passau;Ingolstadt 78;Christopher Keller;76528;4;Kempten 57;1;162;
-24/25;5;n/a;BZL 2 Nord;Passau;Ingolstadt 78;Tim Maier;54554;4;Kempten 57;2;146;
-24/25;5;n/a;BZL 2 Nord;Passau;Ingolstadt 78;Julius Maier;47002;4;Kempten 57;3;131;
-24/25;5;n/a;BZL 2 Nord;Passau;Kempten 57;Stefan Busch;58995;4;Ingolstadt 78;0;207;
-24/25;5;n/a;BZL 2 Nord;Passau;Kempten 57;Ethan Jones;88163;4;Ingolstadt 78;1;145;
-24/25;5;n/a;BZL 2 Nord;Passau;Kempten 57;Oscar Hartmann;78415;4;Ingolstadt 78;2;177;
-24/25;5;n/a;BZL 2 Nord;Passau;Kempten 57;Vincent Smith;49514;4;Ingolstadt 78;3;120;
-24/25;5;n/a;BZL 2 Nord;Passau;Passau 51;Ludwig Scholz;99436;4;Hof 79;0;94;
-24/25;5;n/a;BZL 2 Nord;Passau;Passau 51;Mohammad Wagner;49467;4;Hof 79;1;114;
-24/25;5;n/a;BZL 2 Nord;Passau;Passau 51;William Lehmann;83173;4;Hof 79;2;133;
-24/25;5;n/a;BZL 2 Nord;Passau;Passau 51;Finn Wilson;69639;4;Hof 79;3;177;
-24/25;5;n/a;BZL 2 Nord;Passau;Schweinfurt 23;Marco Lopez;67478;4;Hof 79;0;166;
-24/25;5;n/a;BZL 2 Nord;Passau;Schweinfurt 23;Moritz Sousa;66454;4;Hof 79;1;112;
-24/25;5;n/a;BZL 2 Nord;Passau;Schweinfurt 23;Lars Busch;94819;4;Hof 79;2;190;
-24/25;5;n/a;BZL 2 Nord;Passau;Schweinfurt 23;Ivan Jäger;53814;4;Hof 79;3;164;
